Actualizar Dockerfile_Mywall y script user_data.sh para optimizar la instalación de dependencias y mejorar la copia de archivos

+++ terraform/user_data.sh
@@ -11,8 +11,14 @@ sudo git clone https://github.com/ferminromero00/EKS-SYMFONY.git
 
 # Crear directorio y copiar archivos
 sudo mkdir -p /var/mywall
-sudo cp -r EKS-SYMFONY/dockerfiles/Dockerfile_Mywall /var/mywall/
 sudo cp -r EKS-SYMFONY/Mywall/* /var/mywall/
+sudo cp EKS-SYMFONY/dockerfiles/Dockerfile_Mywall /var/mywall/
+
+# Verificar archivos necesarios
+if [ ! -f /var/mywall/composer.json ]; then
+    echo "Error: composer.json no encontrado"
+    exit 1
+fi
 
 # Instalar Docker
 sudo dnf install -y docker
